+++ src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import { WalletConnectionProvider } from "../context/useWalletConnectionContext"
+import { WalletConnectionProvider } from "../contexts/WalletConnectionProvider"
 import "../styles/globals.css"
 import { ThemeProvider } from "next-themes"
 import type { AppProps } from "next/app"
